优化文件备份功能

+++ FileSystem/TagItemDelegate.cpp
@@ -14,16 +14,12 @@ void TagItemDelegate::paint(QPainter *painter, const QStyleOptionViewItem &optio
 
     if (hasTags(filePath)) {
         QRect iconRect(option.rect.right() - 30, option.rect.top() + 5, 20, 20);
-        // QIcon tagIcon(":/usedimage/edittag.svg");
         QIcon tagIcon = QIcon::fromTheme(QIcon::ThemeIcon::MailMessageNew);
-
         tagIcon.paint(painter, iconRect, Qt::AlignCenter);
     }
     if (dbservice.dbBackup().hasSubmissions(filePath)) {
         QRect submissionIconRect(option.rect.right() - 60, option.rect.top() + 5, 20, 20);
-        // QIcon submissionIcon(":/usedimage/history.svg");
         QIcon submissionIcon = QIcon::fromTheme(QIcon::ThemeIcon::EditCopy);
-
         submissionIcon.paint(painter, submissionIconRect, Qt::AlignCenter);
     }
 }
@@ -113,19 +109,26 @@ void TagItemDelegate::onOpenFileTriggered(QAbstractItemModel *model, const QMode
     emit openFileRequested(filePath);
 }
 
-void TagItemDelegate::onDeleteFileTriggered(QAbstractItemModel *model, const QModelIndex &index) {
-    qDebug() << "TagItemDelegate::onDeleteFileTriggered 文件删除功能隐藏了";
-    return;
 
+void TagItemDelegate::onDeleteFileTriggered(QAbstractItemModel *model, const QModelIndex &index) {
     QString filePath = model->data(index, QFileSystemModel::FilePathRole).toString();
-    QFile file(filePath);
-    if (file.remove()) {
-        qDebug() << "文件已删除:" << filePath;
-        model->removeRow(index.row());
-    } else {
-        qWarning() << "删除文件失败:" << filePath << "错误:" << file.errorString();
+    QMessageBox msgBox;
+    msgBox.setWindowTitle("删除文件");
+    msgBox.setText("你确定要删除这个文件吗?");
+    msgBox.setWindowIcon(QIcon::fromTheme("utilities-system-monitor"));
+    msgBox.setStandardButtons(QMessageBox::Yes | QMessageBox::No);
+    msgBox.setDefaultButton(QMessageBox::No);
+
+    int ret = msgBox.exec();
+    if (ret == QMessageBox::Yes) {
+        QFile file(filePath);
+        if (file.remove()) {
+            model->removeRow(index.row());
+        } else {
+            qWarning() << "删除文件失败:" << filePath << "错误:" << file.errorString();
+        }
+        emit deleteFileRequested(filePath);
     }
-    emit deleteFileRequested(filePath);
 }
 
 void TagItemDelegate::onCommitTriggered(QAbstractItemModel *model, const QModelIndex &index) {
